Convert Windows project to VS2017
The existing code uses an ancient version of Visual Studio (2010), so
this commit updates the project to use the latest version of Visual
Studio.

It seems that VS2017 does not like that the solution and one of the
contained projects have the same name, so I have had to change the
project name to "Liberator".

As part of the conversion it became clear that the few things we used
from Boost had been (mostly) adopted by the C++ standard, so I have
removed the Boost dependency rather than upgrading to a version of
Boost that works the VS2017 C++ compiler.